+++ src/deprecated/routes/RequestsRoute/RequestsRoute.js
@@ -34,6 +34,7 @@ import {
   NoValue,
   MCLPagingTypes,
   Icon,
+  exportToCsv,
 } from '@folio/stripes/components';
 import {
   deparseFilters,
@@ -41,7 +42,6 @@ import {
   SearchAndSort,
 } from '@folio/stripes/smart-components';
 import {
-  exportCsv,
   effectiveCallNumber,
   getHeaderWithCredentials,
 } from '@folio/stripes/util';
@@ -638,7 +638,7 @@ class RequestsRoute extends React.Component {
     this.findResource = this.findResource.bind(this);
     this.toggleModal = this.toggleModal.bind(this);
     this.buildRecords = this.buildRecords.bind(this);
-    // Map to pass into exportCsv
+    // Map to pass into exportToCsv
     this.columnHeadersMap = this.getColumnHeaders(reportHeaders);
     this.expiredHoldsReportColumnHeaders = this.getColumnHeaders(expiredHoldsReportHeaders);
 
@@ -860,7 +860,7 @@ class RequestsRoute extends React.Component {
     this.columnHeadersMap = this.state.isViewPrintDetailsEnabled ? this.columnHeadersMap :
       getFilteredColumnHeadersMap(this.columnHeadersMap);
 
-    exportCsv(recordsToCSV, {
+    exportToCsv(recordsToCSV, {
       onlyFields: this.columnHeadersMap,
       excludeFields: ['id'],
     });
@@ -1211,7 +1211,7 @@ class RequestsRoute extends React.Component {
     }
 
     const recordsToCSV = buildHoldRecords(requests);
-    exportCsv(recordsToCSV, {
+    exportToCsv(recordsToCSV, {
       onlyFields: this.expiredHoldsReportColumnHeaders,
       excludeFields: ['id'],
     });